build: update action to close stale PRs
My original plan of adding a lable to limit those initially
process as outlined in https://github.com/nodejs/node/pull/48051
does not work I think because adding a lable updates the last
update time.

- Removing the need for the lable
- Remove the cron scheduling so that it only runs when I run it manually
- Fix the display name for the action as I missed updating that after
  cut and paste from existing action

The plan will be to find stop dates that should only affect a reasonable
number of PRs at a time and then run in batches using that instead.

Signed-off-by: Michael Dawson <mdawson@devrus.com>

PR-URL: https://github.com/nodejs/node/pull/48196
Reviewed-By: Mestery <mestery@protonmail.com>